backend/requirements.txt
Django==3.2.12
django-filter==2.4.0
djangorestframework_simplejwt==